fix casting issues for temperature. It requires double instead of float in the c# world. in the mssql server world is float. MSSQL float = double C# world

+++ src/AzCloud/AzCloudApp.MessageProcessor.Core.Thermo.DataStore/DataStoreModel/AttendanceDataStore.cs
@@ -18,7 +18,7 @@ namespace AzCloudApp.MessageProcessor.Core.Thermo.DataStore.DataStoreModel
 
         public DateTime? Birth { get; set; }
 
-        public float BodyTemperature { get; set; }
+        public double BodyTemperature { get; set; }
 
         public string CertificateNumber { get; set; }
